archivage et gestion des accès consultants fonctionnels

+++ consultant/consultation.php
@@ -8,13 +8,15 @@
     $id = $_GET['id'];
     require_once '../data.php';
     if(!array_key_exists($id, $other)){
-	    header("Location: ../referent_info.php");
-	    exit;
-    }
+	    $forbidden = 1;
+    }else{
+	    $forbidden = 0;
+		$_SESSION["role"] = "consultant";
+    	$_SESSION["username"] = $id;
+    	$username = $other[$_SESSION["username"]]['user'];
+	}
 
-    $_SESSION["role"] = "consultant";
-    $_SESSION["username"] = $id;
-    $username = $other[$_SESSION["username"]]['user'];
+    
     
     
     if(isset($_POST['deconnexion'])){ // partie pour déconnecter l'utilisateur
@@ -35,7 +37,7 @@
 <body>
 	<table class="bandeau">
 		<tr>
-			<td rowspan="2"><a href="../home.php"><img src="../images/logo.svg"><img></a></td>
+			<td rowspan="2"><a class="rien" href="../home.php"><img src="../images/logo.svg"><img></a></td>
 			<td><h1 id="taille1">JEUNE</h1></td>
 		</tr>
 		<tr>
@@ -60,12 +62,13 @@
 	<br>
     <br>
 
-
-    <h4>informations sur le jeune</h4>
-    nom : <?php echo $users[$username]['name']; ?><br>
-    prénom : <?php echo $users[$username]['firstname']; ?><br>
-    date de naissance : <?php echo $users[$username]['birth']; ?><br>
-    email : <?php echo $users[$username]['email']; ?>
+	<div id="info_jeune">
+    	<h4>informations sur le jeune</h4>
+    	nom : <?php echo $users[$username]['name']; ?><br>
+    	prénom : <?php echo $users[$username]['firstname']; ?><br>
+    	date de naissance : <?php echo $users[$username]['birth']; ?><br>
+    	email : <?php echo $users[$username]['email']; ?>
+	</div>
     <br>
     <br>
     <br>
@@ -73,52 +76,71 @@
     <br>
     <br>
 
-    <div class="liste">
+    <div id="liste">
 		<?php
 			$nbrConfirmedSkill = 0;
 			echo '<br><table><tr>';
 			foreach($other[$id]["skills"] as $number => $id_skill){
 				if($users[$username]["skills"][$number]['id'] == $id_skill){ # vérifier que les expériences enregistrées sont toujours disponibles
                     $skill = $users[$username]["skills"][$number];
-					$nbrConfirmedSkill++;
-					echo '<td class="marge"><h3>' . $skill["environement"] . "</h3><ul><li>description: " . $skill["description"] . "</li><li>début: " . $skill["beginning"] . "</li><li> durée: " . $skill["duration"] . " " . $skill["durationType"] . "</li></ul>";
-					if(!empty($skill['socialSkills'])){
-						echo "<h5>Compétences : savoir-être</h5><ol>";
-						foreach($skill["socialSkills"] as $socialSkill){
-							echo "<li>" . $socialSkill . "</li>";
+					if($users[$username]["skills"][$number]['status'] == "confirmed"){
+						$nbrConfirmedSkill++;
+						echo '<td class="marge"><h3>' . $skill["environement"] . "</h3><ul><li>description: " . $skill["description"] . "</li><li>début: " . $skill["beginning"] . "</li><li> durée: " . $skill["duration"] . " " . $skill["durationType"] . "</li></ul>";
+						if(!empty($skill['socialSkills'])){
+							echo "<h5>Compétences : savoir-être</h5><ol>";
+							foreach($skill["socialSkills"] as $socialSkill){
+								echo "<li>" . $socialSkill . "</li>";
+							}
+							echo "</ol>";
+						}else{
+							echo "<h5>Compétences : savoir-être</h5><br>aucun savoir-être mentionné";
 						}
-						echo "</ol>";
-					}else{
-						echo "<h5>Compétences : savoir-être</h5><br>aucun savoir-être mentionné";
-					}
-					if(!empty($skill['savoir-faire'])){
-						echo "<h5>Compétences : savoir faire</h5><ol>";
-						foreach($skill["savoir-faire"] as $savoir_faire){
-							echo "<li>" . $savoir_faire . "</li>";
+						if(!empty($skill['savoir-faire'])){
+							echo "<h5>Compétences : savoir faire</h5><ol>";
+							foreach($skill["savoir-faire"] as $savoir_faire){
+								echo "<li>" . $savoir_faire . "</li>";
+							}
+							echo "</ol>";
+						}else{
+							echo "<h5>Compétences : savoir faire</h5><br>aucun savoir-faire mentionné";
 						}
-						echo "</ol>";
-					}else{
-						echo "<h5>Compétences : savoir faire</h5><br>aucun savoir-faire mentionné";
+						echo '</td><td class="marge">';
+						echo "<h4>Référent</h4>";
+						echo $skill["referent"]["firstname"] . " " . $skill["referent"]["name"] . "<br>";
+						echo $skill["referent"]["email"] . "<br>";
+						echo $skill["referent"]["situation"] . "<br>";
+						if($skill["comment"] != ""){
+							echo "<br><h5>Commentaire du référent</h5><br>" . $skill["comment"] . "<br>";
+						}
+						echo "</td>";
+						if($nbrConfirmedSkill%1==0){ // nombre de cases max dans une seule ligne
+							echo "</tr><tr>";
+						}
+					}else if($users[$username]["skills"][$number]['status'] == "confirmed"){
+						$nbrArchivedSkill++;
 					}
-					echo '</td><td class="marge">';
-					echo "<h4>Référent</h4>";
-					echo $skill["referent"]["firstname"] . " " . $skill["referent"]["name"] . "<br>";
-					echo $skill["referent"]["email"] . "<br>";
-					echo $skill["referent"]["situation"] . "<br>";
-					if($skill["comment"] != ""){
-						echo "<br><h5>Commentaire du référent</h5><br>" . $skill["comment"] . "<br>";
-					}
-					echo "</td>";
-					if($nbrConfirmedSkill%1==0){ // nombre de cases max dans une seule ligne
-						echo "</tr><tr>";
-					}
+					
 				}
 			}
 			if($nbrConfirmedSkill == 0){
-                echo '<br><br><br><br><br><br><br><br><p>il semblerait qu\'il n\'y ai aucune expérience enregistrée sur ce profil</p>';
+				if($nbrArchivedSkill == 0){
+                	echo '<br><p>il semblerait que ces expériences ne soient plus consultables</p><br><br>';
+				}else{
+					echo '<br><br><br><br><br><br><br><br><p>il semblerait qu\'il n\'y ai aucune expérience consultable enregistrée sur ce profil</p>';
+				}
             }
             echo "</tr></table>";
 		?>
 	</div>
+	<?php 
+		if($forbidden == 1){
+			echo '<script>
+				document.getElementById("liste").innerHTML = "";
+				document.getElementById("info_jeune").innerHTML = "";
+			</script>';
+			echo 'aucune expérience ne semble correspondre, <a href="../consultant_info.php">retrouvez ici les informations relative aux consultants</a>';
+		}
+	?>
+
 </body>
 </html>
